moved non-pages out of pages folder
