+++ components/Start/pages/LoginPage.js
@@ -2,7 +2,7 @@ import { View } from "react-native";
 import _Header from "../elements/_Header";
 import STYLES from "../styles/Styles";
 import TERMS from "../../../settings/Terms";
-import EmailPassword from "./EmailPassword";
+import EmailPassword from "../elements/EmailPassword";
 import { useNavigation } from "@react-navigation/native";
 import _Button from "../elements/_Button";
